# docs/gpt-smcr/11-actions-openapi.yaml
openapi: 3.1.0
info:
  title: SMCR Editorial GPT API
  version: 1.0.0
  description: API de lectura y escritura controlada. Ningún endpoint llama a modelos.
servers:
  - url: https://SaludMentalCostaRica.com/api/gpt/editorial/v1
security:
  - EditorialBearerAuth: []
paths:
...
  /carousels/{carouselId}/versions:
    get:
      operationId: listVersions
      security: [{ EditorialBearerAuth: [editorial:read] }]
      parameters: [{ $ref: '#/components/parameters/CarouselId' }]
      responses: { '200': { description: Versiones } }
    post:
      operationId: createVersion
      security: [{ EditorialBearerAuth: [editorial:write] }]
      parameters:
        - { $ref: '#/components/parameters/CarouselId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/ConfirmationToken' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateVersionRequest' } } }
      responses: { '201': { description: Nueva versión, content: { application/json: { schema: { $ref: '#/components/schemas/MutationResult' } } } } }
...
components:
  securitySchemes:
    EditorialBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  parameters:
...
    CarouselId: { name: carouselId, in: path, required: true, schema: { type: string } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 16 } }
    ConfirmationToken: { name: X-Editorial-Confirmation, in: header, required: false, schema: { type: string } }
  schemas:
...
    CreateVersionRequest:
      type: object
      required: [baseVersionId, spec]
      properties:
        baseVersionId: { type: string }
        spec: { type: object }
        comment: { type: string }
...
    MutationResult:
      type: object
      required: [ok, carouselId, versionId]
      properties:
        ok: { type: boolean }
        carouselId: { type: string }
        versionId: { type: string }
        confirmationRequired: { type: boolean }
        confirmationId: { type: [string, 'null'] }